rename: change folder name provider to global
